LPD-17663 Moves tests to correct module
