+++ modules/apps/asset/asset-categories-test/src/testIntegration/java/com/liferay/asset/categories/change/tracking/test/AssetVocabularyTableReferenceDefinitionTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.asset.change.tracking.test;
+package com.liferay.asset.categories.change.tracking.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.asset.test.util.AssetTestUtil;
