+++ modules/apps/asset/asset-categories-test/src/testIntegration/java/com/liferay/asset/categories/search/test/AssetCategoryIndexerIndexedFieldsTest.java
@@ -3,10 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.asset.search.test;
+package com.liferay.asset.categories.search.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
+import com.liferay.asset.kernel.model.AssetCategory;
 import com.liferay.asset.kernel.model.AssetVocabulary;
+import com.liferay.asset.kernel.service.AssetCategoryService;
 import com.liferay.asset.kernel.service.AssetVocabularyService;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
@@ -41,7 +43,6 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 
-import org.junit.After;
 import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
@@ -54,7 +55,7 @@ import org.junit.runner.RunWith;
  * @author Lucas Marques
  */
 @RunWith(Arquillian.class)
-public class AssetVocabularyIndexerIndexedFieldsTest {
+public class AssetCategoryIndexerIndexedFieldsTest {
 
 	@ClassRule
 	@Rule
@@ -73,22 +74,21 @@ public class AssetVocabularyIndexerIndexedFieldsTest {
 		AssetVocabularyFixture assetVocabularyFixture =
 			new AssetVocabularyFixture(assetVocabularyService, group);
 
+		AssetCategoryFixture assetCategoryFixture = new AssetCategoryFixture(
+			assetCategoryService, assetVocabularyFixture, group);
+
+		_assetCategoryFixture = assetCategoryFixture;
+
+		_assetCategories = assetCategoryFixture.getAssetCategories();
+
 		_assetVocabularies = assetVocabularyFixture.getAssetVocabularies();
-		_assetVocabularyFixture = assetVocabularyFixture;
 
 		_group = group;
 
 		_groups = groupSearchFixture.getGroups();
-
 		_indexedFieldsFixture = new IndexedFieldsFixture(
 			resourcePermissionLocalService, searchEngineHelper, uidFactory,
 			documentBuilderFactory);
-		_defaultLocale = LocaleThreadLocal.getDefaultLocale();
-	}
-
-	@After
-	public void tearDown() throws Exception {
-		LocaleThreadLocal.setDefaultLocale(_defaultLocale);
 	}
 
 	@Test
@@ -97,13 +97,13 @@ public class AssetVocabularyIndexerIndexedFieldsTest {
 
 		setTestLocale(locale);
 
-		AssetVocabulary assetVocabulary =
-			_assetVocabularyFixture.createAssetVocabulary("新しい商品");
+		AssetCategory assetCategory = _assetCategoryFixture.createAssetCategory(
+			"新しい商品");
 
 		String searchTerm = "新しい";
 
 		assertFieldValues(
-			_expectedFieldValues(assetVocabulary), locale, searchTerm);
+			_expectedFieldValues(assetCategory), locale, searchTerm);
 	}
 
 	@Rule
@@ -127,28 +127,31 @@ public class AssetVocabularyIndexerIndexedFieldsTest {
 				).locale(
 					locale
 				).modelIndexerClasses(
-					AssetVocabulary.class
+					AssetCategory.class
 				).queryString(
 					searchTerm
 				).build()));
 	}
 
 	protected void setTestLocale(Locale locale) throws Exception {
-		_assetVocabularyFixture.updateDisplaySettings(locale);
+		_assetCategoryFixture.updateDisplaySettings(locale);
 
 		LocaleThreadLocal.setDefaultLocale(locale);
 	}
+
+	@Inject(
+		filter = "indexer.class.name=com.liferay.asset.kernel.model.AssetCategory"
+	)
+	protected static Indexer<AssetCategory> indexer;
+
+	@Inject
+	protected AssetCategoryService assetCategoryService;
 
 	@Inject
 	protected AssetVocabularyService assetVocabularyService;
 
 	@Inject
 	protected DocumentBuilderFactory documentBuilderFactory;
-
-	@Inject(
-		filter = "indexer.class.name=com.liferay.asset.kernel.model.AssetVocabulary"
-	)
-	protected Indexer<AssetVocabulary> indexer;
 
 	@Inject
 	protected ResourcePermissionLocalService resourcePermissionLocalService;
@@ -168,68 +171,83 @@ public class AssetVocabularyIndexerIndexedFieldsTest {
 	protected UserSearchFixture userSearchFixture;
 
 	private Map<String, String> _expectedFieldValues(
-			AssetVocabulary assetVocabulary)
+			AssetCategory assetCategory)
 		throws Exception {
 
 		Map<String, String> map = HashMapBuilder.put(
+			Field.ASSET_CATEGORY_ID,
+			String.valueOf(assetCategory.getCategoryId())
+		).put(
+			Field.ASSET_CATEGORY_TITLE,
+			StringUtil.lowerCase(assetCategory.getName())
+		).put(
 			Field.ASSET_VOCABULARY_ID,
-			String.valueOf(assetVocabulary.getVocabularyId())
+			String.valueOf(assetCategory.getVocabularyId())
 		).put(
-			Field.COMPANY_ID, String.valueOf(assetVocabulary.getCompanyId())
+			Field.COMPANY_ID, String.valueOf(assetCategory.getCompanyId())
 		).put(
-			Field.ENTRY_CLASS_NAME, AssetVocabulary.class.getName()
+			Field.ENTRY_CLASS_NAME, AssetCategory.class.getName()
 		).put(
-			Field.ENTRY_CLASS_PK,
-			String.valueOf(assetVocabulary.getVocabularyId())
+			Field.ENTRY_CLASS_PK, String.valueOf(assetCategory.getCategoryId())
 		).put(
-			Field.GROUP_ID, String.valueOf(assetVocabulary.getGroupId())
+			Field.GROUP_ID, String.valueOf(assetCategory.getGroupId())
 		).put(
-			Field.NAME, assetVocabulary.getName()
+			Field.NAME, assetCategory.getName()
 		).put(
-			Field.SCOPE_GROUP_ID, String.valueOf(assetVocabulary.getGroupId())
+			Field.SCOPE_GROUP_ID, String.valueOf(assetCategory.getGroupId())
 		).put(
 			Field.STAGING_GROUP, String.valueOf(_group.isStagingGroup())
 		).put(
-			Field.TITLE, assetVocabulary.getName()
+			Field.TITLE, assetCategory.getName()
 		).put(
-			Field.USER_ID, String.valueOf(assetVocabulary.getUserId())
+			Field.USER_ID, String.valueOf(assetCategory.getUserId())
 		).put(
-			Field.USER_NAME, StringUtil.lowerCase(assetVocabulary.getUserName())
+			Field.USER_NAME, StringUtil.lowerCase(assetCategory.getUserName())
 		).put(
-			Field.VISIBILITY_TYPE,
-			String.valueOf(assetVocabulary.getVisibilityType())
+			"assetCategoryTitle_ja_JP",
+			StringUtil.lowerCase(assetCategory.getName())
 		).put(
-			"name_sortable", StringUtil.lowerCase(assetVocabulary.getName())
+			"name_sortable", StringUtil.lowerCase(assetCategory.getName())
 		).put(
-			"title_ja_JP", assetVocabulary.getName()
+			"parentCategoryId",
+			String.valueOf(assetCategory.getParentCategoryId())
 		).put(
-			"title_sortable", StringUtil.lowerCase(assetVocabulary.getName())
+			"title_ja_JP", assetCategory.getName()
 		).put(
-			"visibilityType_sortable",
-			String.valueOf(assetVocabulary.getVisibilityType())
+			"title_sortable", StringUtil.lowerCase(assetCategory.getName())
+		).put(
+			"treePath", assetCategory.getTreePath()
 		).build();
 
-		_indexedFieldsFixture.populateUID(assetVocabulary, map);
+		_indexedFieldsFixture.populateUID(assetCategory, map);
 
-		_populateDates(assetVocabulary, map);
-		_populateRoles(assetVocabulary, map);
-		_populateLocalizedTitles(assetVocabulary.getTitle(), map);
+		_populateDates(assetCategory, map);
+		_populateRoles(assetCategory, map);
+		_populateTitles(assetCategory.getName(), map);
 
 		return map;
 	}
 
 	private void _populateDates(
-		AssetVocabulary assetVocabulary, Map<String, String> map) {
+		AssetCategory assetCategory, Map<String, String> map) {
 
 		_indexedFieldsFixture.populateDate(
-			Field.CREATE_DATE, assetVocabulary.getCreateDate(), map);
+			Field.CREATE_DATE, assetCategory.getCreateDate(), map);
 		_indexedFieldsFixture.populateDate(
-			Field.MODIFIED_DATE, assetVocabulary.getModifiedDate(), map);
+			Field.MODIFIED_DATE, assetCategory.getModifiedDate(), map);
 	}
 
-	private void _populateLocalizedTitles(
-		String title, Map<String, String> map) {
+	private void _populateRoles(
+			AssetCategory assetCategory, Map<String, String> map)
+		throws Exception {
 
+		_indexedFieldsFixture.populateRoleIdFields(
+			assetCategory.getCompanyId(), AssetCategory.class.getName(),
+			assetCategory.getCategoryId(), assetCategory.getGroupId(), null,
+			map);
+	}
+
+	private void _populateTitles(String title, Map<String, String> map) {
 		map.put("localized_title", title);
 
 		for (Locale locale : LanguageUtil.getAvailableLocales()) {
@@ -248,21 +266,14 @@ public class AssetVocabularyIndexerIndexedFieldsTest {
 		}
 	}
 
-	private void _populateRoles(
-			AssetVocabulary assetVocabulary, Map<String, String> map)
-		throws Exception {
+	@DeleteAfterTestRun
+	private List<AssetCategory> _assetCategories;
 
-		_indexedFieldsFixture.populateRoleIdFields(
-			assetVocabulary.getCompanyId(), AssetVocabulary.class.getName(),
-			assetVocabulary.getVocabularyId(), assetVocabulary.getGroupId(),
-			null, map);
-	}
+	private AssetCategoryFixture _assetCategoryFixture;
 
 	@DeleteAfterTestRun
 	private List<AssetVocabulary> _assetVocabularies;
 
-	private AssetVocabularyFixture _assetVocabularyFixture;
-	private Locale _defaultLocale;
 	private Group _group;
 
 	@DeleteAfterTestRun
