+++ modules/apps/asset/asset-categories-test/src/testIntegration/java/com/liferay/asset/categories/search/test/AssetCategoryMultiLanguageSearchTest.java
@@ -3,13 +3,13 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.asset.search.test;
+package com.liferay.asset.categories.search.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.asset.kernel.model.AssetVocabulary;
+import com.liferay.asset.kernel.model.AssetCategory;
+import com.liferay.asset.kernel.service.AssetCategoryService;
 import com.liferay.asset.kernel.service.AssetVocabularyService;
 import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.search.Indexer;
 import com.liferay.portal.kernel.search.SearchEngineHelper;
@@ -19,7 +19,6 @@ import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
 import com.liferay.portal.kernel.test.rule.Sync;
 import com.liferay.portal.kernel.test.rule.SynchronousDestinationTestRule;
-import com.liferay.portal.kernel.test.util.TestPropsValues;
 import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.kernel.util.LocaleThreadLocal;
 import com.liferay.portal.kernel.util.LocaleUtil;
@@ -49,7 +48,7 @@ import org.junit.runner.RunWith;
  */
 @RunWith(Arquillian.class)
 @Sync
-public class AssetVocabularyMultiLanguageSearchTest {
+public class AssetCategoryMultiLanguageSearchTest {
 
 	@ClassRule
 	@Rule
@@ -68,11 +67,16 @@ public class AssetVocabularyMultiLanguageSearchTest {
 		AssetVocabularyFixture assetVocabularyFixture =
 			new AssetVocabularyFixture(assetVocabularyService, group);
 
-		_assetVocabularies = assetVocabularyFixture.getAssetVocabularies();
-		_assetVocabularyFixture = assetVocabularyFixture;
+		AssetCategoryFixture assetCategoryFixture = new AssetCategoryFixture(
+			assetCategoryService, assetVocabularyFixture, group);
+
+		_assetCategories = assetCategoryFixture.getAssetCategories();
+		_assetCategoryFixture = assetCategoryFixture;
 
 		_defaultLocale = LocaleThreadLocal.getDefaultLocale();
+
 		_group = group;
+
 		_groups = groupSearchFixture.getGroups();
 	}
 
@@ -85,25 +89,27 @@ public class AssetVocabularyMultiLanguageSearchTest {
 	public void testEnglishDescription() throws Exception {
 		setTestLocale(LocaleUtil.US);
 
-		_addAssetVocabularyMultiLanguage();
+		_addAssetCategoryMultiLanguage();
+
+		Map<String, String> descriptionMap = HashMapBuilder.put(
+			"description", _ENGLISH_DESCRIPTION
+		).put(
+			"description_en_US", _ENGLISH_DESCRIPTION
+		).put(
+			"description_ja_JP", _JAPANESE_DESCRIPTION
+		).build();
+
+		String keyword = "description";
 
 		assertFieldValues(
-			"description", LocaleUtil.US,
-			HashMapBuilder.put(
-				"description", _ENGLISH_DESCRIPTION
-			).put(
-				"description_en_US", _ENGLISH_DESCRIPTION
-			).put(
-				"description_ja_JP", _JAPANESE_DESCRIPTION
-			).build(),
-			"description");
+			"description", LocaleUtil.US, descriptionMap, keyword);
 	}
 
 	@Test
 	public void testEnglishTitle() throws Exception {
 		setTestLocale(LocaleUtil.US);
 
-		_addAssetVocabularyMultiLanguage();
+		_addAssetCategoryMultiLanguage();
 
 		Map<String, String> titleMap = HashMapBuilder.put(
 			"title", _ENGLISH_TITLE
@@ -116,6 +122,7 @@ public class AssetVocabularyMultiLanguageSearchTest {
 		).build();
 
 		assertFieldValues("title", LocaleUtil.US, titleMap, "title");
+
 		assertFieldValues("title", LocaleUtil.US, titleMap, "tit");
 	}
 
@@ -123,7 +130,7 @@ public class AssetVocabularyMultiLanguageSearchTest {
 	public void testJapaneseDescription() throws Exception {
 		setTestLocale(LocaleUtil.JAPAN);
 
-		_addAssetVocabularyMultiLanguage();
+		_addAssetCategoryMultiLanguage();
 
 		Map<String, String> descriptionMap = HashMapBuilder.put(
 			"description", _JAPANESE_DESCRIPTION
@@ -135,9 +142,12 @@ public class AssetVocabularyMultiLanguageSearchTest {
 
 		assertFieldValues(
 			"description", LocaleUtil.JAPAN, descriptionMap, "新規");
+
 		assertFieldValues(
 			"description", LocaleUtil.JAPAN, descriptionMap, "作成");
+
 		assertFieldValues("description", LocaleUtil.JAPAN, descriptionMap, "新");
+
 		assertFieldValues("description", LocaleUtil.JAPAN, descriptionMap, "作");
 	}
 
@@ -145,7 +155,7 @@ public class AssetVocabularyMultiLanguageSearchTest {
 	public void testJapaneseTitle() throws Exception {
 		setTestLocale(LocaleUtil.JAPAN);
 
-		_addAssetVocabularyMultiLanguage();
+		_addAssetCategoryMultiLanguage();
 
 		Map<String, String> titleMap = HashMapBuilder.put(
 			"title", _JAPANESE_TITLE
@@ -158,8 +168,11 @@ public class AssetVocabularyMultiLanguageSearchTest {
 		).build();
 
 		assertFieldValues("title", LocaleUtil.JAPAN, titleMap, "新規");
+
 		assertFieldValues("title", LocaleUtil.JAPAN, titleMap, "作成");
+
 		assertFieldValues("title", LocaleUtil.JAPAN, titleMap, "新");
+
 		assertFieldValues("title", LocaleUtil.JAPAN, titleMap, "作");
 	}
 
@@ -167,15 +180,15 @@ public class AssetVocabularyMultiLanguageSearchTest {
 	public SearchTestRule searchTestRule = new SearchTestRule();
 
 	protected void assertFieldValues(
-		String prefix, Locale locale, Map<String, String> map,
+		String prefix, Locale locale, Map<String, String> titleStrings,
 		String searchTerm) {
 
 		FieldValuesAssert.assertFieldValues(
-			map, name -> name.startsWith(prefix),
+			titleStrings, name -> name.startsWith(prefix),
 			searcher.search(
 				searchRequestBuilderFactory.builder(
 				).companyId(
-					getCompanyId()
+					_group.getCompanyId()
 				).groupIds(
 					_group.getGroupId()
 				).locale(
@@ -183,34 +196,28 @@ public class AssetVocabularyMultiLanguageSearchTest {
 				).fields(
 					StringPool.STAR
 				).modelIndexerClasses(
-					AssetVocabulary.class
+					AssetCategory.class
 				).queryString(
 					searchTerm
 				).build()));
 	}
 
-	protected long getCompanyId() {
-		try {
-			return TestPropsValues.getCompanyId();
-		}
-		catch (PortalException portalException) {
-			throw new RuntimeException(portalException);
-		}
-	}
-
 	protected void setTestLocale(Locale locale) throws Exception {
-		_assetVocabularyFixture.updateDisplaySettings(locale);
+		_assetCategoryFixture.updateDisplaySettings(locale);
 
 		LocaleThreadLocal.setDefaultLocale(locale);
 	}
 
 	@Inject
+	protected AssetCategoryService assetCategoryService;
+
+	@Inject
 	protected AssetVocabularyService assetVocabularyService;
 
 	@Inject(
-		filter = "indexer.class.name=com.liferay.asset.kernel.model.AssetVocabulary"
+		filter = "indexer.class.name=com.liferay.asset.kernel.model.AssetCategory"
 	)
-	protected Indexer<AssetVocabulary> indexer;
+	protected Indexer<AssetCategory> indexer;
 
 	@Inject
 	protected ResourcePermissionLocalService resourcePermissionLocalService;
@@ -224,8 +231,8 @@ public class AssetVocabularyMultiLanguageSearchTest {
 	@Inject
 	protected SearchRequestBuilderFactory searchRequestBuilderFactory;
 
-	private void _addAssetVocabularyMultiLanguage() throws Exception {
-		_assetVocabularyFixture.createAssetVocabulary(
+	private void _addAssetCategoryMultiLanguage() throws Exception {
+		_assetCategoryFixture.createAssetCategory(
 			new LocalizedValuesMap() {
 				{
 					put(LocaleUtil.US, _ENGLISH_TITLE);
@@ -249,9 +256,9 @@ public class AssetVocabularyMultiLanguageSearchTest {
 	private static final String _JAPANESE_TITLE = "新規作成";
 
 	@DeleteAfterTestRun
-	private List<AssetVocabulary> _assetVocabularies;
+	private List<AssetCategory> _assetCategories;
 
-	private AssetVocabularyFixture _assetVocabularyFixture;
+	private AssetCategoryFixture _assetCategoryFixture;
 	private Locale _defaultLocale;
 	private Group _group;
