+++ modules/apps/asset/asset-categories-test/src/testIntegration/java/com/liferay/asset/categories/search/test/AssetVocabularyMultiLanguageSearchTest.java
@@ -3,13 +3,13 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.asset.search.test;
+package com.liferay.asset.categories.search.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.asset.kernel.model.AssetCategory;
-import com.liferay.asset.kernel.service.AssetCategoryService;
+import com.liferay.asset.kernel.model.AssetVocabulary;
 import com.liferay.asset.kernel.service.AssetVocabularyService;
 import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.search.Indexer;
 import com.liferay.portal.kernel.search.SearchEngineHelper;
@@ -19,6 +19,7 @@ import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
 import com.liferay.portal.kernel.test.rule.Sync;
 import com.liferay.portal.kernel.test.rule.SynchronousDestinationTestRule;
+import com.liferay.portal.kernel.test.util.TestPropsValues;
 import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.kernel.util.LocaleThreadLocal;
 import com.liferay.portal.kernel.util.LocaleUtil;
@@ -48,7 +49,7 @@ import org.junit.runner.RunWith;
  */
 @RunWith(Arquillian.class)
 @Sync
-public class AssetCategoryMultiLanguageSearchTest {
+public class AssetVocabularyMultiLanguageSearchTest {
 
 	@ClassRule
 	@Rule
@@ -67,16 +68,11 @@ public class AssetCategoryMultiLanguageSearchTest {
 		AssetVocabularyFixture assetVocabularyFixture =
 			new AssetVocabularyFixture(assetVocabularyService, group);
 
-		AssetCategoryFixture assetCategoryFixture = new AssetCategoryFixture(
-			assetCategoryService, assetVocabularyFixture, group);
-
-		_assetCategories = assetCategoryFixture.getAssetCategories();
-		_assetCategoryFixture = assetCategoryFixture;
+		_assetVocabularies = assetVocabularyFixture.getAssetVocabularies();
+		_assetVocabularyFixture = assetVocabularyFixture;
 
 		_defaultLocale = LocaleThreadLocal.getDefaultLocale();
-
 		_group = group;
-
 		_groups = groupSearchFixture.getGroups();
 	}
 
@@ -89,27 +85,25 @@ public class AssetCategoryMultiLanguageSearchTest {
 	public void testEnglishDescription() throws Exception {
 		setTestLocale(LocaleUtil.US);
 
-		_addAssetCategoryMultiLanguage();
-
-		Map<String, String> descriptionMap = HashMapBuilder.put(
-			"description", _ENGLISH_DESCRIPTION
-		).put(
-			"description_en_US", _ENGLISH_DESCRIPTION
-		).put(
-			"description_ja_JP", _JAPANESE_DESCRIPTION
-		).build();
-
-		String keyword = "description";
+		_addAssetVocabularyMultiLanguage();
 
 		assertFieldValues(
-			"description", LocaleUtil.US, descriptionMap, keyword);
+			"description", LocaleUtil.US,
+			HashMapBuilder.put(
+				"description", _ENGLISH_DESCRIPTION
+			).put(
+				"description_en_US", _ENGLISH_DESCRIPTION
+			).put(
+				"description_ja_JP", _JAPANESE_DESCRIPTION
+			).build(),
+			"description");
 	}
 
 	@Test
 	public void testEnglishTitle() throws Exception {
 		setTestLocale(LocaleUtil.US);
 
-		_addAssetCategoryMultiLanguage();
+		_addAssetVocabularyMultiLanguage();
 
 		Map<String, String> titleMap = HashMapBuilder.put(
 			"title", _ENGLISH_TITLE
@@ -122,7 +116,6 @@ public class AssetCategoryMultiLanguageSearchTest {
 		).build();
 
 		assertFieldValues("title", LocaleUtil.US, titleMap, "title");
-
 		assertFieldValues("title", LocaleUtil.US, titleMap, "tit");
 	}
 
@@ -130,7 +123,7 @@ public class AssetCategoryMultiLanguageSearchTest {
 	public void testJapaneseDescription() throws Exception {
 		setTestLocale(LocaleUtil.JAPAN);
 
-		_addAssetCategoryMultiLanguage();
+		_addAssetVocabularyMultiLanguage();
 
 		Map<String, String> descriptionMap = HashMapBuilder.put(
 			"description", _JAPANESE_DESCRIPTION
@@ -142,12 +135,9 @@ public class AssetCategoryMultiLanguageSearchTest {
 
 		assertFieldValues(
 			"description", LocaleUtil.JAPAN, descriptionMap, "新規");
-
 		assertFieldValues(
 			"description", LocaleUtil.JAPAN, descriptionMap, "作成");
-
 		assertFieldValues("description", LocaleUtil.JAPAN, descriptionMap, "新");
-
 		assertFieldValues("description", LocaleUtil.JAPAN, descriptionMap, "作");
 	}
 
@@ -155,7 +145,7 @@ public class AssetCategoryMultiLanguageSearchTest {
 	public void testJapaneseTitle() throws Exception {
 		setTestLocale(LocaleUtil.JAPAN);
 
-		_addAssetCategoryMultiLanguage();
+		_addAssetVocabularyMultiLanguage();
 
 		Map<String, String> titleMap = HashMapBuilder.put(
 			"title", _JAPANESE_TITLE
@@ -168,11 +158,8 @@ public class AssetCategoryMultiLanguageSearchTest {
 		).build();
 
 		assertFieldValues("title", LocaleUtil.JAPAN, titleMap, "新規");
-
 		assertFieldValues("title", LocaleUtil.JAPAN, titleMap, "作成");
-
 		assertFieldValues("title", LocaleUtil.JAPAN, titleMap, "新");
-
 		assertFieldValues("title", LocaleUtil.JAPAN, titleMap, "作");
 	}
 
@@ -180,15 +167,15 @@ public class AssetCategoryMultiLanguageSearchTest {
 	public SearchTestRule searchTestRule = new SearchTestRule();
 
 	protected void assertFieldValues(
-		String prefix, Locale locale, Map<String, String> titleStrings,
+		String prefix, Locale locale, Map<String, String> map,
 		String searchTerm) {
 
 		FieldValuesAssert.assertFieldValues(
-			titleStrings, name -> name.startsWith(prefix),
+			map, name -> name.startsWith(prefix),
 			searcher.search(
 				searchRequestBuilderFactory.builder(
 				).companyId(
-					_group.getCompanyId()
+					getCompanyId()
 				).groupIds(
 					_group.getGroupId()
 				).locale(
@@ -196,28 +183,34 @@ public class AssetCategoryMultiLanguageSearchTest {
 				).fields(
 					StringPool.STAR
 				).modelIndexerClasses(
-					AssetCategory.class
+					AssetVocabulary.class
 				).queryString(
 					searchTerm
 				).build()));
 	}
 
+	protected long getCompanyId() {
+		try {
+			return TestPropsValues.getCompanyId();
+		}
+		catch (PortalException portalException) {
+			throw new RuntimeException(portalException);
+		}
+	}
+
 	protected void setTestLocale(Locale locale) throws Exception {
-		_assetCategoryFixture.updateDisplaySettings(locale);
+		_assetVocabularyFixture.updateDisplaySettings(locale);
 
 		LocaleThreadLocal.setDefaultLocale(locale);
 	}
 
 	@Inject
-	protected AssetCategoryService assetCategoryService;
-
-	@Inject
 	protected AssetVocabularyService assetVocabularyService;
 
 	@Inject(
-		filter = "indexer.class.name=com.liferay.asset.kernel.model.AssetCategory"
+		filter = "indexer.class.name=com.liferay.asset.kernel.model.AssetVocabulary"
 	)
-	protected Indexer<AssetCategory> indexer;
+	protected Indexer<AssetVocabulary> indexer;
 
 	@Inject
 	protected ResourcePermissionLocalService resourcePermissionLocalService;
@@ -231,8 +224,8 @@ public class AssetCategoryMultiLanguageSearchTest {
 	@Inject
 	protected SearchRequestBuilderFactory searchRequestBuilderFactory;
 
-	private void _addAssetCategoryMultiLanguage() throws Exception {
-		_assetCategoryFixture.createAssetCategory(
+	private void _addAssetVocabularyMultiLanguage() throws Exception {
+		_assetVocabularyFixture.createAssetVocabulary(
 			new LocalizedValuesMap() {
 				{
 					put(LocaleUtil.US, _ENGLISH_TITLE);
@@ -256,9 +249,9 @@ public class AssetCategoryMultiLanguageSearchTest {
 	private static final String _JAPANESE_TITLE = "新規作成";
 
 	@DeleteAfterTestRun
-	private List<AssetCategory> _assetCategories;
+	private List<AssetVocabulary> _assetVocabularies;
 
-	private AssetCategoryFixture _assetCategoryFixture;
+	private AssetVocabularyFixture _assetVocabularyFixture;
 	private Locale _defaultLocale;
 	private Group _group;
